fetchers/twitter_api/requirements.txt
tweepy==4.14.0
flask==3.0.0
python-dotenv==1.0.0
google-cloud-storage>=2.10.0